Make it possible to update metadata on posts
Since we are now modifying the post yet again after assigning
categories, it's no longer ReplaceCategories' job to refresh the post.

I check for the presence of any arguments to know if the post needs to
be refreshed or not, but it's still pretty stupid (for example, updating
to the exact same metadata or categories don't need to refresh it). This
could be handled via dirty checking, but then we need something to
contain state about the updates.