Expose option to user to delete a post w/o force
The default is not to delete a post with force. When you don't use force the
post just ends up in the trash can where it can be be restored or permanently
deleted.

The is no way via the current WP-API to tell the status of a post. Posts that
are in the trash are still retrivable by the API.