Support native WP API instead of plugin
This client is built to interact with WP REST API plugin.
Wordpress 4.7 has a REST API built in (based on the wp rest plugin), which means that the plugin is not compatible with recent versions of WP.

The native REST API has some more or less subtle differences with the plugin API, especially regarding embedded meta data.